move category expansion types to restmapper package
