+++ pkg/kubectl/cmd/testing/fake.go
@@ -287,7 +287,7 @@ func (f *TestFactory) Cleanup() {
 	os.Remove(f.tempConfigFile.Name())
 }
 
-func (f *TestFactory) CategoryExpander() categories.CategoryExpander {
+func (f *TestFactory) CategoryExpander() restmapper.CategoryExpander {
 	return categories.LegacyCategoryExpander
 }
 
@@ -335,6 +335,7 @@ func (f *TestFactory) Command(*cobra.Command, bool) string {
 
 func (f *TestFactory) NewBuilder() *resource.Builder {
 	mapper, err := f.RESTMapper()
+	categoryExpander, err2 := f.CategoryExpander()
 
 	return resource.NewFakeBuilder(
 		func(version schema.GroupVersion) (resource.RESTClient, error) {
@@ -347,8 +348,8 @@ func (f *TestFactory) NewBuilder() *resource.Builder {
 			return f.Client, nil
 		},
 		mapper,
-		f.CategoryExpander(),
-	).AddError(err)
+		categoryExpander,
+	).AddError(err).AddError(err2)
 }
 
 func (f *TestFactory) KubernetesClientSet() (*kubernetes.Clientset, error) {
